finished be

+++ WashingTime/Controllers/WashingTimeController.cs
@@ -4,9 +4,14 @@ using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Npgsql;
 using WashingTime.Dtos;
 using WashingTime.Dtos.Filters;
 using WashingTime.Entities.WashingTime;
+using WashingTime.Exceptions;
+using WashingTime.Identity;
+using WashingTime.Infrastructure;
 using WashingTime.QueryExtensions.Pagination;
 using WashingTime.Infrastructure.Filters;
 using WashingTimeFilter = WashingTime.Dtos.Filters.WashingTimeFilter;
@@ -20,11 +25,16 @@ namespace WashingTime.Controllers
     {
         private readonly IWashingTimeRepository _repository;
         private readonly IMapper _mapper;
+        private readonly IIdentityAccessor _identityAccessor;
 
-        public WashingTimeController(IWashingTimeRepository repository, IMapper mapper)
+        public WashingTimeController(
+            IWashingTimeRepository repository,
+            IMapper mapper,
+            IIdentityAccessor identityAccessor)
         {
             _repository = repository;
             _mapper = mapper;
+            _identityAccessor = identityAccessor;
         }
 
         [HttpGet]
@@ -43,10 +53,19 @@ namespace WashingTime.Controllers
         public async Task<Entities.WashingTime.WashingTime> Store([FromBody] WashingTimeDto washingTimeDto)
         {
             var helpRequest = _mapper.Map<Entities.WashingTime.WashingTime>(washingTimeDto);
-            helpRequest.UserId =
-                User.FindFirst("http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier").Value;
+            helpRequest.UserId = _identityAccessor.UserId;
             _repository.Add(helpRequest);
-            await _repository.UnitOfWork.SaveEntitiesAsync();
+            try
+            {
+                await _repository.UnitOfWork.SaveEntitiesAsync();
+            }
+            catch (DbUpdateException exception)
+                when (exception.InnerException is PostgresException postgresException &&
+                      postgresException.SqlState == SqlStateCodeStatic.UniqueViolation)
+            {
+                throw new UniqueConstraintViolationException("This time is already taken");
+            }
+
             return helpRequest;
         }
 
@@ -68,6 +87,11 @@ namespace WashingTime.Controllers
         public async Task<Entities.WashingTime.WashingTime> Delete(Guid id)
         {
             var washingTime = await _repository.GetAsync(id);
+            if (washingTime.UserId != _identityAccessor.UserId)
+            {
+                throw new EntityDeletionRestrictionException("You cannot delete another user's washingtime");
+            }
+
             await _repository.DeleteAsync(washingTime);
             await _repository.UnitOfWork.SaveEntitiesAsync();
             return washingTime;
